pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "travel-agenda"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.scripts]
travel = "travel.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
